+++ src/main/java/POT/DuoBloom/question/dto/AnswerDto.java
@@ -1,4 +1,4 @@
-package POT.DuoBloom.feed.dto;
+package POT.DuoBloom.question.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
