+++ src/main/java/POT/DuoBloom/question/dto/QuestionWithAnswersDto.java
@@ -1,18 +1,13 @@
-package POT.DuoBloom.feed.dto;
+package POT.DuoBloom.question.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
 import java.util.List;
 
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
-public class QuestionDto {
+public class QuestionWithAnswersDto {
     private Long questionId;
-    private LocalDate feedDate; // feedDate로 수정
     private String content;
     private String myAnswerStatus;
     private String coupleAnswerStatus;
